+++ qsprpred/data/data.py
@@ -4,8 +4,10 @@ import concurrent
 import copy
 import json
 import os
+import pickle
 import warnings
-from typing import Callable, List, Literal, Union
+from multiprocessing import Pool
+from typing import Callable, List, Literal
 
 import numpy as np
 import pandas as pd
@@ -19,7 +21,7 @@ from qsprpred.data.utils.folds import Folds
 from qsprpred.data.utils.scaffolds import Scaffold
 from qsprpred.data.utils.smiles_standardization import (
     chembl_smi_standardizer,
-    sanitize_smiles,
+    old_sanitize_smiles,
 )
 from qsprpred.logs import logger
 from qsprpred.models.tasks import TargetTasks
@@ -526,20 +528,39 @@ class MoleculeTable(MoleculeDataSet):
         """
         return len([col for col in self.df.columns if col.startswith("ScaffoldGroup_")]) > 0
 
-    def cleanMolecules(self, standardize: bool = True, sanitize: bool = True):
-        """Standardize and or sanitize SMILES sequences."""
-        if standardize:
-            self.standardize()
-        if sanitize:
-            self.sanitize()
+    def standardizeSmiles(self, smiles_standardizer):
+        """Apply smiles_standardizer to the compounds in parallel
 
-    def standardize(self):
-        """Standardize SMILES sequences."""
-        self.df[self.smilescol] = [chembl_smi_standardizer(smiles)[0] for smiles in self.df[self.smilescol]]
+        Args:
+            smiles_standardizer (Union[str, callable]): either `None` to skip the standardization,
+                `chembl`, `old`, or a partial function that reads and standardizes smiles.
 
-    def sanitize(self):
-        """Sanitize SMILES sequences."""
-        self.df[self.smilescol] = [sanitize_smiles(smiles) for smiles in self.df[self.smilescol]]
+        Raises:
+            ValueError: when smiles_standardizer is not a callable or one of the predefined strings.
+        """        
+        std_jobs = self.nJobs
+        if smiles_standardizer is None:
+            return
+        if callable(smiles_standardizer):
+            try: # Prevents weird error if the user inputs a lambda function
+                pickle.dumps(smiles_standardizer)
+            except pickle.PicklingError:
+                logger.warning("Standardizer is not pickleable. Will set n_jobs to 1")
+                std_jobs = 1
+            std_func = smiles_standardizer
+        elif smiles_standardizer.lower() == 'chembl':
+            std_func = chembl_smi_standardizer
+        elif smiles_standardizer.lower() == 'old':
+            std_func = old_standardize_sanitize
+        else:
+            raise ValueError("Standardizer must be either 'chembl', or a callable")
+        
+        if std_jobs == 1:
+            std_smi = [std_func(smi) for smi in self.df[self.smilescol].values]
+        else:
+            with Pool(std_jobs) as pool:
+                std_smi = pool.map(std_func, self.df[self.smilescol].values)
+        self.df[self.smilescol] = std_smi
 
     def shuffle(self, random_state=None):
         """Shuffle the internal data frame."""
@@ -1264,8 +1285,7 @@ class QSPRDataset(MoleculeTable):
 
     def prepareDataset(
         self,
-        standardize=True,
-        sanitize=True,
+        smiles_standardizer='chembl',
         datafilters=None,
         split=None,
         fold=None,
@@ -1278,8 +1298,8 @@ class QSPRDataset(MoleculeTable):
         """Prepare the dataset for use in QSPR model.
 
         Arguments:
-            standardize (bool): Apply Chembl standardization pipeline to smiles
-            sanitize (bool): sanitize smiles
+            smiles_standardizer (Union[str, callable]): either `None` to skip the standardization,
+                `chembl`, `old`, or a partial function that reads and standardizes smiles.
             datafilters (list of datafilter obj): filters number of rows from dataset
             split (datasplitter obj): splits the dataset into train and test set
             fold (datasplitter obj): splits the train set into folds for cross validation
@@ -1291,8 +1311,7 @@ class QSPRDataset(MoleculeTable):
             feature_fill_value (float): value to fill missing values with, defaults to `numpy.nan`
         """
         # apply sanitization and standardization
-        if standardize:
-            self.cleanMolecules(standardize, sanitize)
+        self.standardizeSmiles(smiles_standardizer)
 
         # calculate featureNames
         if feature_calculator is not None: